refactor : Updated UI for Cart page and implemented Total Price feature
